themes: Initial port of our themes for A12
We have a simple color palette to work, so, it won't be 100% accurate but
we can adapt it to the new color palette.
Also, transform AOSP Theme into Unmonet just to respect the Monet
Deniers (tbh not complaining, monet as an accent is cool)

For now, we got 5/11 themes "ported", currently not being
built because of not having an interface for UI Styles.
The following will be deprecated:
- Pissel Theme: Because we're already in the pixel style

Change-Id: Ibf2b2c26aaeb7c4d091abbf33953e5f9eb856e3a