DO NOT MERGE: test for building better overlays
Taking notes from substratum, let's see if we can build overlays
in a way that's not so hacky or doesn't require a lot of exposures

Implementation:
- Using the parameter aapt -I [apkfile] we can ask resources
from an app. So, for example, our systemui got some dimens
for the brightness bar, and instead of hardcoding it every time
it's better to dynamically ask the apk for the values.
That's the way substratum handles the overlay building.

We're setting a weird way to get the final apk and
telling the build system that we're waiting for SystemUI
to be done to then build the overlay to avoid build errors.
The path might not be the best way to do this
but we can maybe get a better approach in the future.

Change-Id: Ifef6afb5a728edba0776be93ced0a5576e978287